use shared seen map
This will allow us to skip processing lines if we already seen them in
another file. While testing this with our tailwindui.com codebase, I saw
the cache hits go from 13,556 to 29,275 hits.

This is an improvement for the initial build in a project with multiple
files. Subsequent rebuilds and projects with a single file won't benefit
from this improvement, but it also won't affect it in a negative way
which is nice.