fix future sort order bug
When sorting based on the BigInt values, we are doing `Number(a - z)`,
but if the result of `a - z` is bigger then the max number, then
overflows would happen. This will result in a hard-to-debug bug.

This is a prevention for the future, because currently all our `a - z`
values will not reach Number.MAX_SAFE_INTEGER mark (9007199254740991).