Support static utilities/components
Unconditionally render rules that we can't reliably turn into "on-demand" plugins. Currently this just checks for the presence of a class, but it should get more sophisticated over time as we think through more situations.

Static utilities/components do not work with variants, and anything that isn't static automatically _does_ work with variants. This may be good or bad, not sure yet. Might want to force people to be explicit about variant support.