implement naive recursive @apply
Currently what it will do is re-run the apply code. If it turns out that
there are @apply rules left, then it will re-apply those. If there are
no @apply rules left, this means that we are "done".

One big downside is that we are (at the time of this commit) running the
apply code for the test at least 4 times. The good part is that we are
doing it on the exact same tree so we don't need to re-parse it.

This is the first implementation which can be improved!